api: hide symbols from public api
