+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/session/SessionIdGenerator.kt
@@ -5,18 +5,6 @@
 
 package io.opentelemetry.android.agent.session
 
-import io.opentelemetry.api.trace.TraceId
-import java.util.Random
-
-interface SessionIdGenerator {
+internal fun interface SessionIdGenerator {
     fun generateSessionId(): String
-
-    object DEFAULT : SessionIdGenerator {
-        override fun generateSessionId(): String {
-            val random = Random()
-            // The OTel TraceId has exactly the same format as a RUM SessionId, so let's re-use it here,
-            // rather than re-inventing the wheel.
-            return TraceId.fromLongs(random.nextLong(), random.nextLong())
-        }
-    }
 }
