+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/session/SessionIdTimeoutHandler.kt
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.android.agent.session
 
+import io.opentelemetry.android.Incubating
 import io.opentelemetry.android.internal.services.applifecycle.ApplicationStateListener
 import io.opentelemetry.sdk.common.Clock
 import kotlin.time.Duration
@@ -34,6 +35,7 @@ internal class SessionIdTimeoutHandler(
     private var state = State.FOREGROUND
 
     // for testing
+    @OptIn(Incubating::class)
     internal constructor(sessionConfig: SessionConfig) : this(
         Clock.getDefault(),
         sessionConfig.backgroundInactivityTimeout,
