+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/session/SessionStorage.kt
@@ -7,18 +7,8 @@ package io.opentelemetry.android.agent.session
 
 import io.opentelemetry.android.session.Session
 
-interface SessionStorage {
+internal interface SessionStorage {
     fun get(): Session
 
     fun save(newSession: Session)
-
-    class InMemory(
-        @Volatile var session: Session = Session.NONE,
-    ) : SessionStorage {
-        override fun get(): Session = session
-
-        override fun save(newSession: Session) {
-            this.session = newSession
-        }
-    }
 }
